Update lists report and add internal list
Previously, the csv from `maintainers report --verbose --details lists`
was populated with the internal list and the maintainer emails. This
change adds the repo name itself to the file. This will enable GitHub
team permissions to be set based on the MAINTAINERS files.

This change also adds a reference to the Google group the Maintainers maintainers are
associated with. By far, this is my most favorite new list name.